# docs/swagger.yml
swagger: '2.0'
info:
  description: 'This is documentation for Tech Community - A Social platform for ICT.'
  version: '1.0.0'
  title: 'Tech Community'
  termsOfService: 'http://swagger.io/terms/'
  contact:
    email: 'email@email.com'
  license:
    name: 'Apache 2.0'
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
host: 'localhost:3000'
basePath: '/api/v1'
...
schemes:
  - 'https'
  - 'http'
consumes:
  - application/json
  - application/xml
produces:
  - application/json
  - application/xml
  - text/html
securityDefinitions:
  bearerAuth:
    type: apiKey
    in: header
    name: Authorization
paths:
...
  /user:
    get:
      summary: 'Get the current user'
      security:
        - bearerAuth: []
      tags:
        - 'Users'
      operationId: getUser
      parameters:
        - name: Authorization
          in: header
          required: true
          type: string
      responses:
        '200':
          description: return current user
          schema:
            $ref: '#/definitions/user'

    put:
      summary: 'Update user profile'
      security:
        - bearerAuth: []
      tags:
        - 'Users'
      operationId: updateUser
      parameters:
        - name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/updateUser'
      responses:
        '200':
          description: returns user information
          schema:
            type: object
            properties:
              user:
                $ref: '#/definitions/user'
...
definitions:
  user:
    type: object
    properties:
      id:
        type: string
        format: uuid
      email:
        type: string
        format: email
      username:
        type: string
      firstName:
        type: string
        example: 'Jake'
      lastName:
        type: string
        example: 'Jake'
      gender:
        type: string
      birthDate:
        type: string
        example: '2019-01-07'
      bio:
        type: string
      image:
        type: string
      cover:
        type: string
      status:
        type: string
        example: 'active'
      token:
        type: string
    required:
      - id
      - email
      - username
      - token
...
  updateUser:
    type: object
    properties:
      user:
        type: object
        properties:
          email:
            type: string
          bio:
            type: string
          image:
            type: string
...